+++ tests/get_name_properties.json
