+++ tests/get_name_properties.py
